backend/requirements.txt
django-cors-headers==3.11.0
Pillow==9.3.0
gunicorn==20.1.0
Django==4.0.1
djangorestframework==3.13.1
djoser==2.2.0